pyproject.toml
[project]
name = "oci-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "Jurgen vijverman",email = "63164608+jurgenvijverman@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "oci (>=2.165.1,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "openai (>=2.15.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
